fix: update path handling

+++ test/bash/cache.bats
@@ -1,24 +1,16 @@
+# shellcheck disable=SC2034,SC2148,SC2155
 
 setup() {
-  load $BATS_SUPPORT_LOAD_PATH
-  load $BATS_ASSERT_LOAD_PATH
+  load "$BATS_SUPPORT_LOAD_PATH"
+  load "$BATS_ASSERT_LOAD_PATH"
 
-    TEST_DIR="$(cd "$(dirname "$BATS_TEST_FILENAME")" >/dev/null 2>&1 && pwd)"
-    TEST_ROOT_DIR=$(mktemp -u)
+  TEST_DIR="$(cd "$(dirname "$BATS_TEST_FILENAME")" >/dev/null 2>&1 && pwd)"
+  TEST_ROOT_DIR=$(mktemp -u)
 
-    load "$TEST_DIR/../../src/usr/local/containerbase/util.sh"
+  load "$TEST_DIR/../../src/usr/local/containerbase/util.sh"
 
-    # load test overwrites
-    load "$TEST_DIR/util.sh"
-
-    # set directories for test
-    ROOT_DIR="${TEST_ROOT_DIR}/root"
-    BIN_DIR="${TEST_ROOT_DIR}/bin"
-    USER_HOME="${TEST_ROOT_DIR}/user"
-    ENV_FILE="${TEST_ROOT_DIR}/env"
-
-    # set default test user
-    TEST_ROOT_USER=1000
+  # load test overwrites
+  load "$TEST_DIR/util.sh"
 }
 
 teardown() {
@@ -27,7 +19,7 @@ teardown() {
 
 @test "get_oldest_file" {
 
-  CONTAINERBASE_CACHE_DIR= \
+  CONTAINERBASE_CACHE_DIR='' \
   run get_oldest_file
   assert_failure
 
@@ -54,7 +46,7 @@ teardown() {
 @test "get_cache_fill_level" {
   local TEST_FILL_LEVEL=88
 
-  CONTAINERBASE_CACHE_DIR= \
+  CONTAINERBASE_CACHE_DIR='' \
   run get_cache_fill_level
   assert_failure
 
@@ -63,11 +55,12 @@ teardown() {
   mkdir -p "${CONTAINERBASE_CACHE_DIR}"
 
   local real_fill_level=$(get_cache_fill_level)
-  assert test "[[ "$real_fill_level" =~ ^[0-9]+$ ]]"
+  assert test "[[ ""$real_fill_level"" =~ ^[0-9]+$ ]]"
 
   # overwrite function to verify deletion
+  # shellcheck disable=SC2317
   function get_cache_fill_level () {
-    echo $TEST_FILL_LEVEL
+    echo "$TEST_FILL_LEVEL"
   }
 
   run get_cache_fill_level
@@ -101,17 +94,17 @@ teardown() {
   sleep 0.01
   touch "${CONTAINERBASE_CACHE_DIR}/a"
 
-  CONTAINERBASE_CACHE_DIR= \
-  CONTAINERBASE_MAX_ALLOCATED_DISK= \
+  CONTAINERBASE_CACHE_DIR='' \
+  CONTAINERBASE_MAX_ALLOCATED_DISK='' \
   run cleanup_cache
   assert_success
 
-  CONTAINERBASE_CACHE_DIR= \
+  CONTAINERBASE_CACHE_DIR='' \
   CONTAINERBASE_MAX_ALLOCATED_DISK=20 \
   run cleanup_cache
   assert_success
 
-  CONTAINERBASE_MAX_ALLOCATED_DISK= \
+  CONTAINERBASE_MAX_ALLOCATED_DISK='' \
   run cleanup_cache
   assert_success
   assert [ -e "${CONTAINERBASE_CACHE_DIR}/a" ]
@@ -162,11 +155,12 @@ teardown() {
   assert_success
   assert_line "${CONTAINERBASE_CACHE_DIR}/foobar"
 
-  CONTAINERBASE_CACHE_DIR= \
+  CONTAINERBASE_CACHE_DIR='' \
   tmp_file=$(download_file "${file}")
   rm "${tmp_file}"
 
-  assert test "[[ "${tmp_file}" =~ "\/containerbase.tar.xz" ]]"
+  # shellcheck disable=SC1001,SC2140
+  assert test "[[ ""${tmp_file}"" =~ "\/containerbase.tar.xz" ]]"
 }
 
 @test "get_from_url" {
@@ -207,7 +201,7 @@ teardown() {
   local checksum="233c335a7f10e9f0dfd7e9d0cda802a38c15a7f13b6678c55980814f22799a70590d56888a819b6591881ec1939240d9dbe68e7e495021b4d6c6a49cdee24d80"
   local file="https://github.com/containerbase/base/releases/download/7.10.0/containerbase.tar.xz"
 
-  run get_from_url "${file}" $(basename "${file}") "${checksum}" "sha512sum"
+  run get_from_url "${file}" "$(basename "${file}")" "${checksum}" "sha512sum"
   assert_success
   assert_line --regexp "^${CONTAINERBASE_CACHE_DIR}/[0-9a-f]{64}/containerbase\.tar\.xz"
 
@@ -220,7 +214,7 @@ teardown() {
   rm -rf "${CONTAINERBASE_CACHE_DIR}"
 
   # wrong checksum
-  run get_from_url "${file}" $(basename "${file}") "123" "sha256sum"
+  run get_from_url "${file}" "$(basename "${file}")" "123" "sha256sum"
   assert_failure
   assert_output --partial "Retries left: 2"
   assert_output --partial "Retries left: 1"
@@ -245,7 +239,7 @@ teardown() {
   local checksum="233c335a7f10e9f0dfd7e9d0cda802a38c15a7f13b6678c55980814f22799a70590d56888a819b6591881ec1939240d9dbe68e7e495021b4d6c6a49cdee24d80"
   local file="https://github.com/containerbase/base/releases/download/7.10.0/containerbase.tar.xz"
 
-  run --separate-stderr get_from_url "${file}" $(basename "${file}") "${checksum}" "sha512sum"
+  run --separate-stderr get_from_url "${file}" "$(basename "${file}")" "${checksum}" "sha512sum"
   assert_success
   assert_output --regexp "^${CONTAINERBASE_CACHE_DIR}/[0-9a-f]{64}/containerbase\.tar\.xz"
 
@@ -259,7 +253,7 @@ teardown() {
   echo "a" >> "${file_path}"
 
   # corrupt file in cache
-  run get_from_url "${file}" $(basename "${file}") "${checksum}" "sha512sum"
+  run get_from_url "${file}" "$(basename "${file}")" "${checksum}" "sha512sum"
   assert_success
   assert_output --partial "Cached file is corrupt"
 }
@@ -294,8 +288,8 @@ teardown() {
   assert_output --regexp "^${CONTAINERBASE_CACHE_DIR}/${path_checksum}/folder\.tar\.zst"
 
   # delete cache entry
-  rm -rf ${CONTAINERBASE_CACHE_DIR}/${key_checksum}
-  rm -rf ${CONTAINERBASE_CACHE_DIR}/${path_checksum}
+  rm -rf "${CONTAINERBASE_CACHE_DIR:?}/${key_checksum}"
+  rm -rf "${CONTAINERBASE_CACHE_DIR:?}/${path_checksum}"
 }
 
 @test "restore_folder_from_cache" {
@@ -358,6 +352,6 @@ teardown() {
   assert [ -e "${tool_folder}/b" ]
 
   # delete cache entry
-  rm -rf ${CONTAINERBASE_CACHE_DIR}/${key_checksum}
-  rm -rf ${CONTAINERBASE_CACHE_DIR}/${path_checksum}
+  rm -rf "${CONTAINERBASE_CACHE_DIR:?}/${key_checksum}"
+  rm -rf "${CONTAINERBASE_CACHE_DIR:?}/${path_checksum}"
 }
